auth_service/requirements.txt
fastapi
uvicorn[standard]
python-jose
passlib[bcrypt]
SQLAlchemy
psycopg2-binary
alembic
pytest
httpx
